+++ src/plugins/event.py
@@ -44,6 +44,7 @@ async def getEvent(request: Request, response: Response, authorization: str = He
         cur.execute(f"SELECT eventid, tmplink, departure, destination, distance, mts, dts, img, title, attendee, vote, pvt FROM event WHERE eventid = {eventid} {limit}")
         t = cur.fetchall()
         if len(t) == 0:
+            response.status_code = 404
             return {"error": True, "descriptor": "Event not found"}
         tt = t[0]
         attendee = tt[9].split(",")
@@ -211,6 +212,7 @@ async def eventVote(request: Request, response: Response, authorization: str = H
     cur.execute(f"SELECT vote FROM event WHERE eventid = {eventid}")
     t = cur.fetchall()
     if len(t) == 0:
+        response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "event_not_found")}
     vote = t[0][0].split(",")
     if str(userid) in vote:
@@ -298,6 +300,7 @@ async def patchEvent(request: Request, response: Response, authorization: str = 
     cur.execute(f"SELECT userid FROM event WHERE eventid = {eventid}")
     t = cur.fetchall()
     if len(t) == 0:
+        response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "event_not_found")}
     creator = t[0][0]
     
@@ -327,6 +330,7 @@ async def deleteEvent(eventid: int, request: Request, response: Response, author
     cur.execute(f"SELECT * FROM event WHERE eventid = {eventid}")
     t = cur.fetchall()
     if len(t) == 0:
+        response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "event_not_found")}
     
     cur.execute(f"UPDATE event SET eventid = -eventid WHERE eventid = {eventid}")
@@ -361,6 +365,7 @@ async def updateEventAttendee(request: Request, response: Response, authorizatio
     cur.execute(f"SELECT attendee FROM event WHERE eventid = {eventid}")
     t = cur.fetchall()
     if len(t) == 0:
+        response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "event_not_found")}
     orgattendees = t[0][0].split(",")
     while "" in orgattendees: